Fix the ArrayIndexOutOfBoundsException of ByteBucketArray
The framework starts to use a preserved staging-public-group for new
attributes on the new platform. The group has a bigger bucket id whose
leading bit is 1, which causes a Java signed/unsigned number problem in
multiple byte-integer conversions.

The bucket id has four bits which has a maximum number 15(0xf). However,
when the leading bit is 1, Java will auto-cast the byte to a negative
integer and cause an ArrayIndexOutOfBoundsException.

To solve this, converting the id to an unsigned integer can help Java
treat it as a positive number.

Fixes: 197494608
Bug: 196297712
Test: robotest
Change-Id: Ie5027621df2bf134ca5c5cee60c423db47f47ca7